utf8: use size_t
These were rebased onto upstream patches that include an
overflow-checking cast. We should get rid of that cast now.

+++ utf8.c
@@ -206,7 +206,7 @@ int utf8_width(const char **start, size_t *remainder_p)
  * string, assuming that the string is utf8.  Returns strlen() instead
  * if the string does not look like a valid utf8 string.
  */
-int utf8_strnwidth(const char *string, size_t len, int skip_ansi)
+size_t utf8_strnwidth(const char *string, size_t len, int skip_ansi)
 {
 	const char *orig = string;
 	size_t width = 0;
@@ -231,7 +231,7 @@ int utf8_strnwidth(const char *string, size_t len, int skip_ansi)
 	return cast_size_t_to_int(string ? width : len);
 }
 
-int utf8_strwidth(const char *string)
+size_t utf8_strwidth(const char *string)
 {
 	return utf8_strnwidth(string, strlen(string), 0);
 }
